Removed old ApiKey references and rename ApiToken to ApiKey
