+++ test/trento_web/auth/api_key_test.exs
@@ -1,80 +1,63 @@
 defmodule TrentoWeb.ApiKeyTest do
-  use ExUnit.Case
-  use Trento.DataCase
-
-  alias Trento.Settings
+  @moduledoc false
 
   alias TrentoWeb.Auth.ApiKey
 
-  setup_all do
-    %{
-      scenarios: [
-        %{
-          installation_id: "some-installation-id",
-          otp25_api_key:
-            "SFMyNTY.g2gDdAAAAAFkAA9pbnN0YWxsYXRpb25faWRtAAAAFHNvbWUtaW5zdGFsbGF0aW9uLWlkYQBkAAhpbmZpbml0eQ.1BK6v9h8jw3JVavFvYWjp90PZpOGtLcnys6mSgqcLQM",
-          api_key:
-            "SFMyNTY.g2gDdAAAAAF3D2luc3RhbGxhdGlvbl9pZG0AAAAUc29tZS1pbnN0YWxsYXRpb24taWRhAHcIaW5maW5pdHk.MjjWMw7hUQ0e8__Ivt21HfQk9mLZ-PfI0DhXqSVbnFA"
-        },
-        %{
-          installation_id: "another-installation-id",
-          otp25_api_key:
-            "SFMyNTY.g2gDdAAAAAFkAA9pbnN0YWxsYXRpb25faWRtAAAAF2Fub3RoZXItaW5zdGFsbGF0aW9uLWlkYQBkAAhpbmZpbml0eQ.acX8VYHlDyG3lDBiDKgl02nfw-hHnY4gk3040Lxna4c",
-          api_key:
-            "SFMyNTY.g2gDdAAAAAF3D2luc3RhbGxhdGlvbl9pZG0AAAAXYW5vdGhlci1pbnN0YWxsYXRpb24taWRhAHcIaW5maW5pdHk.gM7LWxBCpWoreWSlWCcCxEMyo39uD3ikVdeobJ-nTmg"
-        }
-      ]
-    }
+  use ExUnit.Case
+
+  import Mox
+
+  setup [:set_mox_from_context, :verify_on_exit!]
+
+  @test_timestamp 1_671_641_814
+
+  setup do
+    expect(
+      Joken.CurrentTime.Mock,
+      :current_time,
+      2,
+      fn ->
+        @test_timestamp
+      end
+    )
+
+    :ok
   end
 
-  describe "Signing data and generating an API key" do
-    test "should always generate expected api_key for current installation", %{
-      scenarios: scenarios
-    } do
-      Enum.each(scenarios, fn %{installation_id: installation_id, api_key: expected_api_key} ->
-        Enum.each(1..3, fn _ ->
-          assert expected_api_key ==
-                   ApiKey.sign(%{
-                     installation_id: installation_id
-                   })
-        end)
-      end)
-    end
-  end
+  describe "generate_api_key!/2" do
+    test "should generate and sign a jwt token with the default claims and expiration correctly set" do
+      expected_expiry = @test_timestamp + 400
 
-  describe "Verifying an API key" do
-    test "should reject an unauthenticated API key" do
-      assert {:error, :unauthenticated} = ApiKey.verify("definitely-a-fake-api-key")
-    end
+      expiry_datetime = DateTime.from_unix!(expected_expiry)
 
-    test "should verify an authenticated API key encrypted with erlang OTP25 version", %{
-      scenarios: scenarios
-    } do
-      Enum.each(scenarios, fn %{installation_id: installation_id, otp25_api_key: api_key} ->
-        assert {:ok,
-                %{
-                  installation_id: ^installation_id
-                }} = ApiKey.verify(api_key)
-      end)
+      token = ApiKey.generate_api_key!(%{}, expiry_datetime)
+      {:ok, claims} = Joken.peek_claims(token)
+
+      assert %{
+               "iss" => "https://github.com/trento-project/web",
+               "aud" => "trento_api_key",
+               "exp" => ^expected_expiry,
+               "iat" => @test_timestamp,
+               "jti" => _,
+               "nbf" => @test_timestamp,
+               "typ" => "Bearer"
+             } = claims
     end
 
-    test "should verify an authenticated API key", %{scenarios: scenarios} do
-      Enum.each(scenarios, fn %{installation_id: installation_id, api_key: api_key} ->
-        assert {:ok,
-                %{
-                  installation_id: ^installation_id
-                }} = ApiKey.verify(api_key)
-      end)
-    end
-  end
+    test "should merge the custom claims with the default after signing" do
+      token =
+        ApiKey.generate_api_key!(
+          %{
+            "sub" => 1
+          },
+          DateTime.from_unix!(@test_timestamp)
+        )
 
-  describe "Getting the API key" do
-    test "should provide the API key of the current installation" do
-      installation_id = Settings.get_installation_id()
-      api_key = ApiKey.get_api_key()
+      {:ok, claims} = Joken.peek_claims(token)
 
-      assert {:ok, decoded_data} = ApiKey.verify(api_key)
-      assert %{installation_id: ^installation_id} = decoded_data
+      assert %{
+               "sub" => 1
+             } = claims
     end
   end
 end
